Implement git utility functions and enhance e2e task generation logic to skip when no changes are detected in the spruce directory.
