+++ .evergreen/scripts/generate-tasks.js
@@ -1,4 +1,3 @@
-import { execSync } from "child_process";
 import { writeFileSync } from "fs";
 import { join, parse } from "path";
 import {
@@ -6,42 +5,9 @@ import {
   IGNORED_FILE_EXTENSIONS,
   TASK_MAPPING,
 } from "./constants.js";
+import { whatChanged } from "./git-utils.js";
 
 // This file is written in plain JS because it makes the generator super fast. No need to install TypeScript.
-
-const getMergeBase = () => {
-  try {
-    const mergeBaseCmd = execSync("git merge-base main@{upstream} HEAD")
-      .toString()
-      .trim();
-    return mergeBaseCmd;
-  } catch (e) {
-    throw new Error("getting merge-base", { cause: e });
-  }
-};
-
-/**
- * whatChanged returns a list of files modified in this patch or PR.
- * Prior art from Evergreen:
- * https://github.com/evergreen-ci/evergreen/blob/ab7d4112b352b759acd54c685524177018467c30/cmd/generate-lint/generate-lint.go#L30
- * @returns a string array of modified file names relative to the git root directory
- */
-const whatChanged = () => {
-  const mergeBase = getMergeBase();
-  try {
-    const diffFiles = execSync(`git diff ${mergeBase} --name-only`)
-      .toString()
-      .trim();
-
-    // If there is no diff, this is not a patch build.
-    if (!diffFiles) {
-      return [];
-    }
-    return diffFiles.split("\n").map((file) => file.trim());
-  } catch (e) {
-    throw new Error("getting diff", { cause: e });
-  }
-};
 
 /**
  * targetsFromChangedFiles returns a list of build variants to run based on a list of changed files.
@@ -84,10 +50,12 @@ const generateTasks = () => {
     changes.length === 0
       ? Object.keys(TASK_MAPPING)
       : targetsFromChangedFiles(changes);
+  
   const buildvariants = targets.map((bv) => ({
     name: bv,
     tasks: TASK_MAPPING[bv].map((name) => ({ name })),
   }));
+
   return { buildvariants };
 };
